bunny: splitDomain breaks delegated sub-domain
The splitDomain algorithm assumes the domain is of "length" 2 and
disregards where the SOA is found. This might have been the only
way this provider could work in the past, but at this time it seems
bunny responds with a correct SOA-RR.

Virtual ( pseudo) sub-domain with A-RR continue to function.